functions/recognizer/requirements.txt
numpy
Pillow
opencv-python-headless
tensorflow==2.12.0
fastapi
uvicorn